Test cases for invalid types and type conversions respectively
These test cases are not reasonable input, but they are here to ensure
that we handle these degenerate cases in the same way as Terraform itself
would handle it, at least in terms of whether or not an error is returned.

These cases also serve to ensure we don't have any crasher bugs from this
sort of nonsense.