Looser parsing of variable type specifications
The initial implementation here was based on an earlier iteration of
the HCL2-based config loader in Terraform, which was preserving the prior
single-keyword-based type specifications.

The latest implementation now allows full type expressions, like
list(string) to require a list with string members. Rather than
implementing a type expression evaluator here, and thus risking that
future extensions to this format would generate errors, we instead just
take whatever syntactically-valid expression the user provides and return
it verbatim as a string.

This means that invalid type specifications won't generate errors, but
the primary goal of this library is to extract interface information and
so looser validation is a necessary compromise to meet our goal of being
forwards-compatible with future Terraform changes.