go.mod: Upgrade to HCL v2.20.1
The latest version of the Terraform requires a newer version of HCL than
was previously selected here, because it makes use of the new possibility
for namespaced functions.

This upgrade should make it now possible for programs that use this library
to analyze Terraform modules that contain calls to namespaced functions.